+++ wfx_fmac_driver/bus/sl_wfx_bus.c
@@ -234,7 +234,7 @@ static sl_status_t sl_wfx_bus_access(sl_wfx_host_bus_transfer_type_t type,
       && (type == SL_WFX_BUS_READ)
       && (address != SL_WFX_CONTROL_REG_ID)
       && (sl_wfx_context->used_buffers <= 1)
-      && ((*((uint16_t *)((uint8_t *)buffer + length - SL_WFX_CTRL_REGISTER_SIZE)) & SL_WFX_CONT_NEXT_LEN_MASK) == 0)) {
+      && ((*((uint16_t *)((uint8_t *)buffer + length - SL_WFX_CONT_REGISTER_SIZE)) & SL_WFX_CONT_NEXT_LEN_MASK) == 0)) {
     /* Ask the host opinion on whether the WFx should be put back to sleep or
        not*/
     if (sl_wfx_host_sleep_grant(type, address, length) == SL_STATUS_WIFI_SLEEP_GRANTED) {