+++ wfx_fmac_driver/sl_wfx_constants.h
@@ -18,6 +18,7 @@
 #define SL_WFX_CONSTANTS_H
 
 #include "sl_status.h"
+#include "sl_wfx_api.h"
 #include <stdint.h>
 
 /******************************************************
@@ -176,51 +177,25 @@ static inline uint32_t uint32_identity(uint32_t x)
 #define SL_WFX_DEBUG_TX_REG  0x0200
 #define SL_WFX_DEBUG_FW_LOAD 0x0400
 
-#define IE_RSNE_ID                48
-#define IE_RSNE_CIPHER_SUITE_TKIP 0x02AC0F00
-#define IE_RSNE_CIPHER_SUITE_CCMP 0x04AC0F00
-
-#define IE_VENDOR_SPECIFIC_ID 221
-#define IE_WPA_OUI            0x0050F2
-
-#define PDS_HF_CLK_KEY            'e'
-#define PDS_POWER_CONFIG_KEY      'h'
-#define PDS_ANTENNA_SEL_KEY       'j'
-
 #define PDS_KEY_A       'a'
 #define PDS_KEY_B       'b'
 #define PDS_KEY_C       'c'
 #define PDS_KEY_D       'd'
 #define PDS_KEY_E       'e'
 #define PDS_KEY_F       'f'
+#define PDS_ANTENNA_SEL_KEY       'j'
 
-#define WFX_PTE_INFO              0x0900C0C0
-#define PTE_INFO_KEYSET_IDX       0x0D
-#define PTE_INFO_SIZE             0x10
-
+#define SL_WFX_PTE_INFO                  0x0900C0C0
 #define SL_WFX_MSG_ID_GENERAL_API_MASK   0x20
-
 #define SL_WFX_MSG_INFO_INTERFACE_OFFSET 1
 #define SL_WFX_MSG_INFO_INTERFACE_MASK   0x06
+#define SL_WFX_OPN_SIZE                  14
+
 #ifdef SL_WFX_USE_SECURE_LINK
 #define SL_WFX_MSG_INFO_SECURE_LINK_OFFSET 6
 #define SL_WFX_MSG_INFO_SECURE_LINK_MASK   0xC0
-#endif
-
-#define SL_WFX_CONT_FRAME_TYPE_OFFSET    14
-
-#define SL_WFX_CONFIG_REVISION_OFFSET    24
-#define SL_WFX_CONFIG_REVISION_MASK      0x7
-#define SL_WFX_CONFIG_TYPE_OFFSET        31
-#define SL_WFX_CONFIG_TYPE_MASK          0x1
-
-#define SL_WFX_OPN_SIZE                  14
-
-#define SL_WFX_CTRL_REGISTER_SIZE        (2)
-
-#ifdef SL_WFX_USE_SECURE_LINK
 /* Secure link constants*/
-#define SECURE_LINK_MAC_KEY_LENGTH      32
+#define SECURE_LINK_MAC_KEY_LENGTH         32
 
 #define SL_WFX_SECURE_LINK_SESSION_KEY_LENGTH          (16)
 
@@ -398,6 +373,7 @@ typedef struct {
   uint8_t  encryption_bitmap[SL_WFX_SECURE_LINK_ENCRYPTION_BITMAP_SIZE];
   uint8_t  secure_link_session_key[SL_WFX_SECURE_LINK_SESSION_KEY_LENGTH];
   uint8_t  secure_link_renegotiation_state;
+  sl_wfx_securelink_exchange_pub_keys_ind_t  secure_link_exchange_ind;
 #endif //SL_WFX_USE_SECURE_LINK
 } sl_wfx_context_t;